+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### New Features in This Release
-
+## New Features in This Release
+- feat(Application): add replacing separators by comma
 - feat(Application): add catching custom separator
 - feat: setup project